pyproject.toml
[project]
name = "sorrydb"
version = "0.1.0"
description = "Build a continously updating database of sorry statments in public Lean4 repositories"
authors = [
	{ name = "Lenny Taelman", email = "l.d.j.taelman@uva.nl" },
	{ name = "Austin Letson", email = "waustinletson@gmail.com" },
	{ name = "Oliver Dressler", email = "odressler@gmail.com" },
]
license = "Apache-2.0"
readme = "README.md"
requires-python = ">=3.12,<3.14"


[tool.poetry.dependencies]
gitpython = "^3.1.44"
typer = "^0.15.3"
pydantic-settings = "^2.9.1"
requests = "^2.32.3"
langchain = "^0.3.20"
langchain-anthropic = "^0.3.10"
dotenv = "^0.9.9"
langchain-openai = "^0.3.8"
langchain-google-genai = "^2.1.0"
fastapi = {extras = ["standard"], version = "^0.115.12"}


[tool.poetry.group.dev.dependencies]
pytest = "^8.3.4"
pre-commit = "^4.2.0"
ruff = "^0.11.1"

[tool.poetry.group.orchestration.dependencies]
prefect = {version = "^3.4.1", extras = ["docker"]}

[tool.poetry]
packages = [{ include = "sorrydb" }]


[tool.poetry.scripts]
sorrydb = "sorrydb.cli.sorrydb_cli:app"
run_rfl_agent = "sorrydb.cli.run_rfl_agent:main"
run_llm_agent = "sorrydb.cli.run_llm_agent:main"
deploy_sorrydb = "orchestration.deploy:app"


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.ruff.lint]
select = ["I"]

[tool.pytest.ini_options]
# do not retain tmp_paths because tests create large `.lake` folders which we want to remove after every test
tmp_path_retention_policy = "none"
